fix: Fix bug refresh token expire when sending request
- Before: If token expired right before sending request then backend will response invalid token
- After: Add a tiny delay to token expire earlier real expire time for backend not response invalid token if token expire while sending request

+++ frontend/src/app/core/interceptors/api.interceptor.ts
@@ -73,8 +73,10 @@ export const apiInterceptor: HttpInterceptorFn = (req, next) => {
         return next(clonedRequest);
       }),
       catchError((error) => {
-        authService.deleteToken();
-        router.navigate(['/auth/login']);
+        if (error.status === 401) {
+          authService.deleteToken();
+          router.navigate(['/auth/login']);
+        }
         return throwError(() => error);
       }),
     );
